refactor: JPA용 트랜잭션매니저 생성
